Part three of doorman PR (see #54) (#82)
Builds on #81

A general purpose ID allocator. The allocator gives out small integer identifiers, along with a reference-counted reservation that holds them as "in use".

IDs are eagerly recycled once the reference count hits zero.

Useful for Xania and doorman to both have a "vote" for which channels are in-use. Previously we had bugs where there was a race in doorman if someone disconnected while xania was about to send a ton of info while another connection came in. The new connection would be given the channel ID of the recently-closed connection and would thus be delivered all the messages for the old connection.