Add a "player()" function.
Many of the newer routines handled switched mobs by checking multiple
conditions. This extracts a "player()" function on the CHAR_DATA,
which returns the CHAR_DATA of the "real player" behind a character,
be they the character themselves, a controlling switched IMM, or
nullptr for a genuine, uncontrolled-by-player-character NPC.

I think this is equivalent to the earlier code and reduces complexity.
I wonder if the command interpreter can subsume this at some point,
with commands tagged as "no, this has to occur on the actual PC, not
the switched mob they might be controlling", to save further, though
that might be unobvious.