CHAR_DATA graduates to an actual class.
Not renaming yet, but soon. It now has a constructor and destructor. Naked new()s and delete()s abound, but
that's a problem for another day. The "owner" is a tricky concept to get over. This _might_ be cause for a shared_ptr...with
things like the descriptor having a part share, as well as the "room" a CHAR is in.

PCs are now lumped in with active mob count. Could be fixed, but really...