Defer CHAR_DATA deletion.
Defer deletion of characters until we know none of the game logic
is running. This allows code like stuff doing damage to a person
can assume the victim of the damage is valid even if they die as
a result of the hit, at least until all the characters have been
processed.

Should fix a couple of read-after-free that @snellers caught when
running locally.